fix: operation schema circular import error
